+++ src/tomate/data_base.py
@@ -582,44 +582,53 @@ class DataBase():
         """
         self.data = self.allocate(shape)
 
-    def set_data(self, variable: str, data: np.ndarray):
+    def set_data(self, variable: str, data: np.ndarray,
+                 keyring: Keyring = None):
         """Set the data for a single variable.
 
         :param var: Variable to set the data to.
         :param data: Array of the correct shape for currently
             selected coordinates. Has no axis for variables.
+        :param keyring: [opt] If no data is loaded, loaded scope
+            is fetched from available scope with this keyring.
+            'var' key has no effect.
 
         :raises KeyError: If the variable is not in available scope.
         :raises IndexError: If the data has not the right dimension.
         :raises ValueError: If the data is not of the shape of loaded scope.
         """
+        def check_shape(data):
+            if self.acs.shape(data)[1:] != self.shape[1:]:
+                raise ValueError("data of wrong shape ({}, expected {})"
+                                .format(self.acs.shape(data), self.shape[1:]))
+
         if variable not in self.avail:
             raise KeyError(f"{variable} is not in avail scope. Use add_variable.")
         if self.acs.ndim(data) != self.ncoord:
             raise IndexError("data of wrong dimension ({}, expected {})"
                              .format(data.ndim, self.ncoord))
-        if self.acs.shape(data) != self.shape[1:]:
-            raise ValueError("data of wrong shape ({}, expected {})"
-                             .format(self.acs.shape(data), self.shape[1:]))
 
         data = np.expand_dims(data, 0)
 
         # No data is loaded
         if self.loaded.is_empty():
-            self.loaded = self.avail.copy()
-            self.loaded.slice(var=variable)
+            self.loaded = self.get_subscope('avail', keyring=keyring, var=variable)
+            check_shape(data)
             self.data = data
 
         # Variable is already loaded
         elif variable in self.loaded.var:
+            check_shape(data)
             self[variable][:] = data[0]
 
         # Variable is not loaded, others are
         else:
             self.loaded.var.append(variable)
+            check_shape(data)
             self.data = self.acs.concatenate((self.data, data), axis=0)
 
-    def add_variable(self, variable: str, data: np.ndarray = None, **attrs: Any):
+    def add_variable(self, variable: str, data: np.ndarray = None,
+                     keyring: Keyring = None, **attrs: Any):
         """Add new variable.
 
         Add variable to available scope,
@@ -629,6 +638,9 @@ class DataBase():
         :param variable: Variable to add.
         :param data: [opt] Corresponding data to add.
             Its shape must match that of the loaded scope.
+        :param keyring: [opt] If there is data to set and no data is loaded,
+            loaded scope is fetched from available scope with this keyring.
+            'var' key has no effect.
         :param attrs: [opt] Variable attributes.
             Passed to VariablesInfo.add_variable
         """
@@ -637,7 +649,7 @@ class DataBase():
         else:
             self.avail.var.append(variable)
         if data is not None:
-            self.set_data(variable, data)
+            self.set_data(variable, data, keyring=keyring)
         self.vi.set_attrs(variable, **attrs)
 
     def remove_loaded_variable(self, variables: Union[str, List[str]]):